1/requirements.txt
Pillow==8.2.0
pytesseract==0.3.7
selenium==3.141.0
urllib3==1.26.4
